+++ packages/core/test/helpers/producers.ts
@@ -46,6 +46,9 @@ export const createClearTextReporterOptions = factoryMethod<ClearTextReporterOpt
   allowEmojis: false,
   logTests: true,
   maxTestsToLog: 3,
+  reportTests: true,
+  reportMutants: true,
+  reportScoreTable: true,
 }));
 
 export type ConcurrencyTokenProviderMock = sinon.SinonStubbedInstance<I<ConcurrencyTokenProvider>> & {