+++ packages/instrumenter/testResources/instrumenter/console-sample.js
@@ -1,0 +1,3 @@
+console.log(a + b);
+
+consolee.log(a + b);