+++ packages/util/src/flat-map.ts
@@ -1,7 +1,0 @@
-import flatMap from 'lodash.flatmap';
-
-/**
- * `Array.prototype.flatMap`, but also works on node 10
- * @see https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Array/flatMap
- */
-export { flatMap };