+++ packages/instrumenter/CHANGELOG.md
@@ -3,6 +3,16 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [7.2.0](https://github.com/stryker-mutator/stryker-js/compare/v7.1.1...v7.2.0) (2023-10-02)
+
+### Bug Fixes
+
+- **npm package:** ignore unused files ([#4405](https://github.com/stryker-mutator/stryker-js/issues/4405)) ([f14e789](https://github.com/stryker-mutator/stryker-js/commit/f14e78944652ceccd205ca1541465292e758c565))
+
+### Features
+
+- **string mutations:** don't mutate Symbol descriptions ([#4407](https://github.com/stryker-mutator/stryker-js/issues/4407)) ([bdd0d5c](https://github.com/stryker-mutator/stryker-js/commit/bdd0d5c96c51371f347c06d66555ff255aaf3a6e))
+
 ## [7.1.1](https://github.com/stryker-mutator/stryker-js/compare/v7.1.0...v7.1.1) (2023-07-15)
 
 **Note:** Version bump only for package @stryker-mutator/instrumenter